Terminados iterativo, recursivo y pila

+++ algoritmos_lab_core/src/main/java/vaporware/practica1/Combinatorio.java
@@ -1,14 +1,16 @@
 package vaporware.practica1;
 
+import java.math.BigInteger;
+
 public abstract class Combinatorio {
     public Combinatorio(){
 
     }
 
-    public long calcularCombinatorio(int n, int k){
+    public BigInteger calcularCombinatorio(int n, int k){
         return algoritmo(n,k);
     }
 
-    protected abstract long algoritmo(int n, int k);
+    protected abstract BigInteger algoritmo(int n, int k);
     public abstract String getTipo();
 }
